backend/reviews/serializers.py
from rest_framework import serializers
from .models import Review, Rating

class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Review
        fields = ['id', 'user', 'listing', 'content', 'created_at']

class RatingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Rating
        fields = ['id', 'review', 'score']
