backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
python-dotenv>=0.20,<1.0
psycopg2-binary>=2.9,<3.0
